chore: move driver command types to separate file
